# Vectorized the GMPEs for Europe