--- a/openquake/hazardlib/gsim/atkinson_boore_2003.py
+++ b/openquake/hazardlib/gsim/atkinson_boore_2003.py
@@ -219,7 +219,7 @@ class AtkinsonBoore2003SInter(GMPE):
 
     kind = 'SInter'
 
-    def compute(self, ctx, imts, mean, sig, tau, phi):
+    def compute(self, ctx: np.recarray, imts, mean, sig, tau, phi):
         """
         See :meth:`superclass method
         <.base.GroundShakingIntensityModel.compute>`
@@ -304,7 +304,7 @@ class AtkinsonBoore2003SSlab(AtkinsonBoore2003SInter):
     #: Supported tectonic region type is subduction interface
     DEFINED_FOR_TECTONIC_REGION_TYPE = const.TRT.SUBDUCTION_INTRASLAB
 
-    def compute(self, ctx, imts, mean, sig, tau, phi):
+    def compute(self, ctx: np.recarray, imts, mean, sig, tau, phi):
         """
         See :meth:`superclass method
         <.base.GroundShakingIntensityModel.compute>`