--- a/openquake/hazardlib/gsim/cauzzi_faccioli_2008.py
+++ b/openquake/hazardlib/gsim/cauzzi_faccioli_2008.py
@@ -163,7 +163,7 @@ class CauzziFaccioli2008(GMPE):
 
     kind = "2008"
 
-    def compute(self, ctx, imts, mean, sig, tau, phi):
+    def compute(self, ctx: np.recarray, imts, mean, sig, tau, phi):
         """
         See :meth:`superclass method
         <.base.GroundShakingIntensityModel.compute>`