--- a/openquake/hazardlib/gsim/faccioli_cauzzi_2006.py
+++ b/openquake/hazardlib/gsim/faccioli_cauzzi_2006.py
@@ -53,7 +53,7 @@ class FaccioliCauzzi2006(GMPE):
 
     REQUIRES_DISTANCES = {'repi'}
 
-    def compute(self, ctx, imts, mean, sig, tau, phi):
+    def compute(self, ctx: np.recarray, imts, mean, sig, tau, phi):
         """
         See :meth:`superclass method
         <.base.GroundShakingIntensityModel.compute>`