--- a/openquake/hazardlib/gsim/toro_2002.py
+++ b/openquake/hazardlib/gsim/toro_2002.py
@@ -105,7 +105,7 @@ class ToroEtAl2002(GMPE):
     #: no fault style adjustement in the base class
     CONSTS_FS = {}
 
-    def compute(self, ctx, imts, mean, sig, tau, phi):
+    def compute(self, ctx: np.recarray, imts, mean, sig, tau, phi):
         """
         See :meth:`superclass method
         <.base.GroundShakingIntensityModel.compute>`