--- a/openquake/hazardlib/gsim/youngs_1997.py
+++ b/openquake/hazardlib/gsim/youngs_1997.py
@@ -52,12 +52,16 @@ def _compute_mean(C, A1, A2, A3, A4, A5, A6, mag, hypo_depth,
     Compute mean for subduction interface events, as explained in table 2,
     page 67.
     """
-    mag = mag[idx]
-    rrup = rrup[idx]
-    hypo_depth = hypo_depth[idx]
+    if isinstance(mag, np.ndarray):
+        mag = mag[idx]
+        hypo_depth = hypo_depth[idx]
     mean[idx] = (A1 + A2 * mag + C['C1'] + C['C2'] * (A3 - mag) ** 3 +
-                 C['C3'] * np.log(rrup + A4 * np.exp(A5 * mag)) +
+                 C['C3'] * np.log(rrup[idx] + A4 * np.exp(A5 * mag)) +
                  A6 * hypo_depth)
+
+
+def get(array, idx):
+    return array[idx] if isinstance(array, np.ndarray) else array
 
 
 class YoungsEtAl1997SInter(GMPE):
@@ -131,7 +135,7 @@ class YoungsEtAl1997SInter(GMPE):
                               CONSTS['A6_rock'], ctx.mag, ctx.hypo_depth,
                               ctx.rrup, mean[m], idx_rock)
                 sig[m, idx_rock] += C['C4'] + C['C5'] * np.clip(
-                    ctx.mag[idx_rock], 0, 8.)
+                    get(ctx.mag, idx_rock), 0, 8.)
 
                 if imt == SA(period=4.0, damping=5.0):
                     mean[m] /= 0.399
@@ -144,7 +148,7 @@ class YoungsEtAl1997SInter(GMPE):
                               CONSTS['A6_soil'], ctx.mag, ctx.hypo_depth,
                               ctx.rrup, mean[m], idx_soil)
                 sig[m, idx_soil] += C['C4'] + C['C5'] * np.clip(
-                    ctx.mag[idx_soil], 0, 8.)
+                    get(ctx.mag, idx_soil), 0, 8.)
 
             if (self.DEFINED_FOR_TECTONIC_REGION_TYPE ==
                     const.TRT.SUBDUCTION_INTRASLAB):  # sslab correction