Part of https://github.com/gem/oq-engine/issues/6973. This gives a huge speedup (over 4x) in "computing mean_std". Here are the figures on the spot machine:
```
# master
| calc_21, maxmem=73.7 GB | time_sec | memory_mb | counts    |
|-------------------------+----------+-----------+-----------|
| total classical         | 65_426   | 321.6     | 628       |
| computing mean_std      | 31_275   | 0.0       | 19_457    |
| get_poes                | 14_501   | 0.0       | 19_457    |
| composing pnes          | 7_555    | 0.0       | 9_099_194 |
| make_contexts           | 7_390    | 0.0       | 9_160_616 |

# this branch
| calc_22, maxmem=86.2 GB | time_sec | memory_mb | counts    |
|-------------------------+----------+-----------+-----------|
| total classical         | 43_612   | 367.8     | 628       |
| get_poes                | 16_086   | 0.0       | 19_457    |
| composing pnes          | 7_842    | 0.0       | 9_099_194 |
| make_contexts           | 7_542    | 0.0       | 9_160_616 |
| computing mean_std      | 7_264    | 0.0       | 19_457    |
```